+++ app/src/main/java/com/damasahhre/hooftrim/database/models/FarmWithNextVisit.java
@@ -3,7 +3,7 @@ package com.damasahhre.hooftrim.database.models;
 import com.damasahhre.hooftrim.models.MyDate;
 
 public class FarmWithNextVisit {
-    public Integer farmId;
+    public Long farmId;
     public MyDate nextVisit;
