+++ app/src/main/java/com/damasahhre/hooftrim/database/models/NextReport.java
@@ -4,7 +4,7 @@ import com.damasahhre.hooftrim.models.MyDate;
 
 public class NextReport {
     public MyDate nextVisitDate;
-    public Integer cowId;
+    public Long cowId;
     public Integer cowNumber;
     public String farmName;
 }
